Change to parentheses
This way we don't need to translate and create different strings for the "on", which will appear odd in languages other than English. I also added it for FIO and OpenAlias.

Further, I removed the 1 instance of openalias_alert_title. It can be removed from the localization file if/when it's convenient.

I did NOT change anything for Unstoppable, because that appeared to touch a dependency. That has its own custom (but similar) text response